split out `TxMeta`
