+++ lib/core/cardano-wallet-core.cabal
@@ -276,6 +276,7 @@ library
       Cardano.Wallet.Primitive.Types.Tx.CBOR.Hash
       Cardano.Wallet.Primitive.Types.Tx.TransactionInfo
       Cardano.Wallet.Primitive.Types.Tx.Tx
+      Cardano.Wallet.Primitive.Types.Tx.TxMeta
       Cardano.Wallet.Primitive.Types.UTxO
       Cardano.Wallet.Primitive.Types.UTxOIndex
       Cardano.Wallet.Primitive.Types.UTxOIndex.Internal
