+++ lib/core/src/Cardano/Wallet/Primitive/Types/Tx.hs
@@ -101,7 +101,7 @@ module Cardano.Wallet.Primitive.Types.Tx
 
     ) where
 
-
 import Cardano.Wallet.Primitive.Types.Tx.CBOR
-import Cardano.Wallet.Primitive.Types.Tx.Tx
 import Cardano.Wallet.Primitive.Types.Tx.TransactionInfo
+import Cardano.Wallet.Primitive.Types.Tx.Tx
+import Cardano.Wallet.Primitive.Types.Tx.TxMeta
